added lines to check whether fimo_out exists and if it does, delete it

+++ src/motif_distance.py
@@ -3,7 +3,8 @@ __author__ = 'Jonathan Rubin'
 import os
 
 def run(ranked_file,filedir,MEMEDB,DATABASE,SINGLEMOTIF):
-    os.system("rm -r " + filedir + "fimo_out/")
+    if os.path.isdir(filedir + "fimo_out/"):
+        os.system("rm -r " + filedir + "fimo_out/")
     if SINGLEMOTIF == False:
         os.system("fimo -o " + filedir + "fimo_out/ " + MEMEDB + DATABASE + " " + ranked_file)
     else:
